Suggestion for missing fields in the RSS Feed
Hi, looking at the API, there is information in the DB that is not present in the RSS feed. Would it be possible to add it?

* ```dateEventStart```
* ```dateEventEnd```
* ```location```

To impact the current logic as little as possible, I'd suggest extending the ```description```:
* from 'CfP for %3$s runs from %1$s to %2$s' 
* to 'CfP for %3$s runs from %1$s to %2$s. The event runs from %4$s to %5$s in %6$s'

Thanks,
Tim

PS: If none of those changes are possible, could we at least add the ```eventUri``` used to compute the SHA-1 hash of the entry? That way I could retrieve the missing data via a GET Request on the REST API. You could use the ```setId``` method of Zend for that. Thanks!